Moved spectral db parsers to util.spectraldb
