+++ src/main/java/net/sf/mzmine/modules/peaklistmethods/io/spectraldbsubmit/GnpsJsonGenerator.java
@@ -34,8 +34,8 @@ import javax.json.JsonArray;
 import javax.json.JsonArrayBuilder;
 import javax.json.JsonObjectBuilder;
 import net.sf.mzmine.datamodel.DataPoint;
-import net.sf.mzmine.modules.peaklistmethods.identification.spectraldbsearch.dbentry.DBEntryField;
 import net.sf.mzmine.parameters.Parameter;
+import net.sf.mzmine.util.spectraldb.entry.DBEntryField;
 
 /**
  * Json for GNPS library entry submission
