+++ src/main/java/net/sf/mzmine/util/spectraldb/entry/SpectralDBEntry.java
@@ -16,7 +16,7 @@
  * USA
  */
 
-package net.sf.mzmine.modules.peaklistmethods.identification.spectraldbsearch.dbentry;
+package net.sf.mzmine.util.spectraldb.entry;
 
 import java.util.Map;
 import java.util.Optional;
